requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
browser-cookie3==0.20.1
bs4==0.0.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
cssselect==1.3.0
curl_cffi==0.12.0
distro==1.9.0
fake-useragent==2.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.1.3
frozenlist==1.7.0
future==1.0.0
gallery_dl==1.30.2
greenlet==3.2.3
h11==0.16.0
idna==3.10
imageio-ffmpeg==0.6.0
instaloader==4.14.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
lz4==4.4.4
MarkupSafe==3.0.3
multidict==6.6.3
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pillow==12.1.1
playwright==1.54.0
propcache==0.3.2
py-notifier==0.5.0
pyaes==1.6.1
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pyee==13.0.0
PyExecJS==1.5.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pyquery==2.0.1
Pyrogram==2.0.106
PySocks==1.7.1
pystyle==2.9
python-dotenv==1.1.1
pywin32==311
requests==2.32.4
selenium==4.34.2
setuptools==80.9.0
shadowcopy==0.0.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
termcolor==3.1.0
tiktok-dlpy==2.2.1
tls-client==1.0.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
uv==0.9.11
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.6
WinToaster==0.1.0
WMI==1.5.1
wsproto==1.2.0
yarl==1.20.1
youtube-dl==2021.12.17
yt-dlp==2025.12.8
